requirements.txt
APScheduler==3.6.3
asgiref==3.6.0
cachetools==4.2.2
certifi==2022.9.24
pip==21.3.1
psycopg2==2.9.5
python-telegram-bot==13.14
pytz-deprecation-shim==0.1.0.post0
pytz==2022.6
requests==2.28.1
setuptools==60.2.0
six==1.16.0
sqlparse==0.4.3
tornado==6.1
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.13
wheel==0.37.1
